Add bare bones implementation of PPV.
Starting out the focus is on Pillars and Pipelines.
Actions through Objectives are in a workable state, with a simple Action
Zone Dashboard.
There's a lot missing for Alignment, Daily Tracking is integrated with
Daily Notes, and Yearly template is setup for planning. Everything
in the middle isn't yet setup, Weeks, Months, Quarters. Those are up
next as these time periods eventuate.